refactor(accounts): nest settings endpoint under Users (plural) namespace
Reviewer asked why a top-level Account module rather than under User.
Align to the established plural users API surface: every current-user
endpoint is /api/v1/users/me/* (contact_details, school_name, ...), so
the read lives there too.

- route namespace :account -> :users; URL is now GET /api/v1/users/me/settings
- Api::V1::Account::SettingsController -> Api::V1::Users::SettingsController
- serializer is User::SettingsSerializer (nested under the User model,
  matching User::DataRetentionStatus et al.)

13 tests, 86 assertions, 0 failures. RuboCop + zeitwerk:check clean.
